backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# AWS SDK
boto3==1.29.7
botocore==1.32.7

# Task scheduling
celery==5.3.4
redis==5.0.1

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Billing
stripe==7.4.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
structlog==23.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
